+++ src/main/java/model/board/element/deposable/Bomb.java
@@ -41,7 +41,7 @@ public class Bomb extends Entity {
      * @param owner   The player who owns the bomb.
      * @param board   The game board where the bomb exists.
      */
-    public Bomb(int x, int y, int width, int height, double velocity, int range, Image image, boolean alive, boolean visible, Player owner, Board board) {
+    public Bomb(double x, double y, int width, int height, double velocity, int range, Image image, boolean alive, boolean visible, Player owner, Board board) {
         super(x, y, BOMB_WIDTH.getSize(), BOMB_HEIGHT.getSize(), velocity, image, alive, visible);
         this.owner = owner;
         this.board = board;
