+++ src/main/java/model/board/element/powerup/handicap/SlowDownBonus.java
@@ -6,7 +6,7 @@ import model.board.element.powerup.BonusWithTimer;
 import java.awt.*;
 
 public class SlowDownBonus extends BonusWithTimer {
-    public SlowDownBonus(int x, int y, int width, int height, double velocity, Image image, boolean alive, boolean visible, Player owner) {
+    public SlowDownBonus(double x, double y, int width, int height, double velocity, Image image, boolean alive, boolean visible, Player owner) {
         super(x, y, width, height, velocity, image, alive, visible, owner);
     }
