+++ src/main/java/model/board/element/powerup/handicap/SmallerRangeBonus.java
@@ -6,7 +6,7 @@ import model.board.element.powerup.BonusWithTimer;
 import java.awt.*;
 
 public class SmallerRangeBonus extends BonusWithTimer {
-    public SmallerRangeBonus(int x, int y, int width, int height, double velocity, Image image, boolean alive, boolean visible, Player owner) {
+    public SmallerRangeBonus(double x, double y, int width, int height, double velocity, Image image, boolean alive, boolean visible, Player owner) {
         super(x, y, width, height, velocity, image, alive, visible, owner);
     }
